Fixed names and code error

+++ app/src/main/java/br/com/teste/sicredi/feature/detail/repository/mapper/EvenDetailMapper.kt
@@ -22,5 +22,5 @@ object EvenDetailMapper {
         }
     }
 
-    fun mapCheckIn(payload: CheckInResponse) = CheckIn(payload.code)
+    fun mapCheckIn(payload: CheckInResponse) = CheckIn(payload.code.toInt())
 }
